Spaceships-Unity-VR-Project
Virtual reality game in which the user can interact with the environment
without the need for any physical controllers. Built using a sensor
mounted on the front of the headset to track the hands. The game was
built in Unity engine (scripted in C#), using the Oculus Rift CV1
headset and the Leap Motion sensor. The game involves the player
positioned in the cockpit of the spaceship, which they drive by
interacting with the controls using their own hands.